+++ src/lib/utils/string.ts
